Route WWMath's rotation helpers through reference-semantics Sin/Cos
With the Inv_Sqrt fix the idle Casino recording matches the PC at all eleven
checkpoints and the driving one matches through frame 300, where before it
diverged at 100. It still diverges from 400, when its dozer starts to drive and
turn.

The reference is 32-bit MSVC built /arch:SSE2, so ordinary float and double
arithmetic is IEEE SSE on both sides, and every other __asm in Core is dead code
(#if 0, VC6-only or __ICL-only). What remains is the CRT: 32-bit MSVC has no float
transcendentals, so sinf(x) there is an inline (float)sin((double)x), while
bionic's sinf is its own single-precision algorithm and differs in the last bit on
about 1.2% of angles. matrix3d.h, matrix3.h and vector3.h call sinf/cosf directly
in every rotation helper, and PhysicsUpdate's Rotate_X/Y/Z and the locomotor's
In_Place_Pre_Rotate_Z run them every frame for a moving unit -- so a turning
vehicle is expected to leave the PC within a few dozen frames, which is the
window between 300 and 400.

All twenty sites now use WWMath::Sin/Cos, which evaluate in double and round once;
TanTrig uses (float)tan((double)x); DynamicShroudClearingRangeUpdate uses the
trig.h Sin/Cos. Every float transcendental reachable from logic now goes through
one of those or Trig.cpp; the other hits are diagnostics, Intel-compiler-only code
and network latency maths.

Unlike the Inv_Sqrt fix this cannot be checked against a frozen value before
building, because the effect compounds along the path; the recording decides.

+++ Core/Libraries/Source/WWVegas/WWMath/wwmath.h
@@ -190,13 +190,15 @@ static WWINLINE float CosTrig(float x)
 #endif
 }
 
+// GeneralsX @bugfix Android port 22/09/2026 tanf -> (float)tan((double)x), the reference
+// client's tanf on 32-bit MSVC; see the note on WWMath::Sin/Cos.
 static WWINLINE float TanTrig(float x) 
 { 
 #ifdef USE_DETERMINISTIC_MATH
 	// TODO: return GameMath::Tan(x);
-	return tanf(x); 
+	return (float)tan((double)x); 
 #else
-	return tanf(x); 
+	return (float)tan((double)x); 
 #endif
 }
